server/requirements.txt
dnspython~=2.7.0
idna~=3.10
httpcore~=1.0.9
httpx~=0.28.1
cryptography~=44.0.2
anyio~=4.9.0
h11~=0.16.0
pip~=25.1
PySocks~=1.7.1
cffi~=1.17.1
rich~=14.0.0
Pygments~=2.19.1
markdown-it-py~=3.0.0
PyYAML~=6.0.2
sniffio~=1.3.1
pytest~=8.3.5
typing_extensions~=4.13.2
click~=8.1.8
zstandard~=0.23.0
mdurl~=0.1.2
typer~=0.15.3
shellingham~=1.5.4
MarkupSafe~=3.0.2
pluggy~=1.5.0
iniconfig~=2.1.0
certifi~=2025.1.31
fastapi~=0.115.12
starlette~=0.46.2
pydantic~=2.11.4
email_validator~=2.2.0
pydantic_core~=2.33.2
requests~=2.32.3
websockets~=15.0.1
urllib3~=2.4.0
uvicorn~=0.34.2
httptools~=0.6.4
watchfiles~=1.0.5
annotated-types~=0.7.0
charset-normalizer~=3.4.1
packaging~=25.0
pycparser~=2.22
Jinja2~=3.1.6
python-multipart~=0.0.20
psycopg2-binary>=2.9.10
ntplib~=0.4.0
pytest-cov~=6.1.1
black~=24.4.2
mypy~=1.10.0
types-requests
numpy~=2.2.5
#checkstyle/lint rules
flake8~=7.0.0
flake8-bugbear~=24.2.6
flake8-comprehensions~=3.14.0
flake8-import-order~=0.18.2
flake8-docstrings~=1.7.0
pydocstyle~=6.3.0
slowapi~=0.1.9
types-requests
sqlalchemy~=2.0.41
python-dotenv
psycopg[binary,pool]
